payment

+++ frontend/src/app/home/cart/cart.component.ts
@@ -1,13 +1,14 @@
-import { Component, OnInit } from '@angular/core';
-import { CartService } from '../../service/cart-service.service';
-import { jwtDecode } from 'jwt-decode';
-import { Cart } from '../../model/cart.model';
-import { Product } from '../../model/product.model';
-import { FormsModule } from '@angular/forms';
-import { ImageService } from '../../service/img-service.service';
-import { CommonModule } from '@angular/common';
+import {Component, OnInit} from '@angular/core';
+import {CartService} from '../../service/cart-service.service';
+import {jwtDecode} from 'jwt-decode';
+import {Cart} from '../../model/cart.model';
+import {Product} from '../../model/product.model';
+import {FormsModule} from '@angular/forms';
+import {ImageService} from '../../service/img-service.service';
+import {CommonModule} from '@angular/common';
 import {VNPayService} from "../../service/payment-service.service";
 import {Router, RouterLink} from "@angular/router";
+import {SharedDataService} from "../../service/shared-data.service";
 
 @Component({
   selector: 'app-cart',
@@ -20,17 +21,21 @@ export class CartComponent implements OnInit {
 
   constructor(private cartService: CartService,
               private imgService: ImageService,
-              private paymentService:VNPayService,
-              private router:Router) { }
+              private paymentService: VNPayService,
+              private sharedDataService: SharedDataService,
+              private router: Router) {
+  }
 
   cart: Cart[] = [];
   user_Id!: string;
   product!: number;
-  totalPrice!:number;
+  totalPrice!: number;
   item: any;
-  urlPayment!:string;
+  urlPayment!: string;
   maxQuantity: number = 10;
   imgAvatars: { [key: string]: string } = {};
+  carts!: Cart;
+  productID: any;
 
   ngOnInit(): void {
     const token = localStorage.getItem('auth_token');
@@ -41,33 +46,53 @@ export class CartComponent implements OnInit {
     this.getAllByUserId();
   }
 
-  payment(totalPrice:number,productName:any){
-    this.paymentService.submitOrder(totalPrice,productName).subscribe((data:any) =>{
-      this.urlPayment = data.vnpayUrl;
-      window.location.href = this.urlPayment;
-    })
+
+  payment() {
+    if (this.selectedProducts.length === 0) {
+      alert('Vui lòng chọn ít nhất một sản phẩm để thanh toán!');
+      return;
+    }
+
+    const totalAmount = this.selectedProducts.reduce((sum, cart) => sum + (cart.product_quantity * cart.product.price), 0);
+    console.log(totalAmount);
+
+    this.paymentService.submitOrder(totalAmount, 'Sản phẩm đã chọn').subscribe((data: any) => {
+      const urlPayment = data.vnpayUrl;
+      window.location.href = urlPayment;
+      localStorage.setItem('myArray', JSON.stringify({
+        user: this.selectedProducts[0].user,
+        products: this.selectedProducts.map(cart => {
+          return ({
+            product: cart.product
+          });
+        })
+      }));
+
+    });
   }
 
+
   getAllByUserId() {
-      this.cartService.getByUserId(this.user_Id).subscribe((data: any) => {
-        this.cart = data.result;
-        this.cart.forEach((cart) => {
-          this.getImageFromService(cart.product.image, cart.cart_id)
-        })
+    this.cartService.getByUserId(this.user_Id).subscribe((data: any) => {
+      this.cart = data.result;
+      this.cart.forEach((cart) => {
+        this.getImageFromService(cart.product.image, cart.cart_id)
       })
+    })
   }
 
   private getImageFromService(imageName: any, product_id: string): void {
     if (imageName) {
       this.imgService.getImage(imageName).subscribe(
         (data: any) => {
-          const blob = new Blob([data], { type: 'image/*' });
+          const blob = new Blob([data], {type: 'image/*'});
           this.imgAvatars[product_id] = URL.createObjectURL(blob);
         },
         error => {
           console.log(error);
         });
-    } else { }
+    } else {
+    }
   }
 
 
@@ -85,12 +110,25 @@ export class CartComponent implements OnInit {
     }
   }
 
-  deleteCart(cart_id:string){
-    console.log(cart_id)
-    this.cartService.delete(cart_id).subscribe(() => this.getAllByUserId());
+  deleteCart(cart_id: string) {
+    if (window.confirm("Are you sure you want to delete this cart")) {
+      console.log(cart_id)
+      this.cartService.delete(cart_id).subscribe(() => this.getAllByUserId());
+    }
   }
 
   updateCart(carts: Cart) {
-    this.cartService.updateCart(carts.cart_id, carts).subscribe(data => { });
+    this.cartService.updateCart(carts.cart_id, carts).subscribe(data => {
+    });
+  }
+
+  selectedProducts: any[] = [];
+
+  toggleSelection(cart: any): void {
+    if (cart.selected) {
+      this.selectedProducts.push(cart);
+    } else {
+      this.selectedProducts = this.selectedProducts.filter(item => item.cart_id !== cart.cart_id); // Bỏ chọn
+    }
   }
 }
